+++ example/lib/main.dart
@@ -1,12 +1,20 @@
 import 'package:flutter/material.dart';
-import 'package:event_crew/event_crew.dart';
+
+import 'home.dart';
 
 void main() {
   runApp(const MyApp());
 }
 
-class MyApp extends StatelessWidget {
+class MyApp extends StatefulWidget {
+
   const MyApp({super.key});
+
+  @override
+  State<MyApp> createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
 
   // This widget is the root of your application.
   @override
@@ -25,7 +33,7 @@ class MyApp extends StatelessWidget {
         // is not restarted.
         primarySwatch: Colors.blue,
       ),
-      home: Home()
+      home: HomePage(),
     );
   }
 }