+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## 0.0.6
+
+* Fix sound path
+* Dynamic lotte animtion file
+
+## 0.0.5
+
+* Dynamic setState between onTap & OnChange Page
+
+## 0.0.4
+
+* Dynamic PageView & List Item
+
+## 0.0.3
+
+* Manipulate package
+* Export all dart code
+
+## 0.0.2
+
+* Integrate existing code into package
+
 ## 0.0.1
 
-Describe initial release.
+* Initial release.